pyproject.toml
[project]
name = "simple-gs"
version = "0.1.0"
description = "Simple Ground Station for UBC Rocket - decodes COBS/CRC/Protobuf serial packets"
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
    "cobs>=1.2.1",
    "crcmod>=1.7",
    "protobuf>=5.29.0",
]

[project.scripts]
simple-gs = "simple_gs.decoder:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
